Re-factored to move context-related integrity checks below Context namespace.
